requirements.txt
fastapi
uvicorn
pandas
azure-ai-documentintelligence
azure-core
